requirements.txt
torch==2.1.0
torchvision>=0.16.0
numpy>=1.24.0
opencv-python>=4.8.0
librosa>=0.10.0
scikit-learn>=1.3.0
streamlit>=1.25.0
streamlit
torch
torchvision
opencv-python
numpy
pandas
librosa
scikit-learn
transformers
Pillow